gh52 - updated module names and added some docstrings
